+++ src/libs/systems/ProgramExiter.ts
@@ -1,4 +1,4 @@
-import { AppError, ErrorReasons, reportError } from '../../models/ExitReasons';
+import { AppError, ErrorReasons, reportError } from '../../models/ErrorReasons';
 import { errorLog } from '../Log';
 
 /**
